requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
attrs=22.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.0=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py38haa95532_0
cffi=1.15.1=py38h2bbff1b_3
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=py38haa95532_0
comm=0.1.3=pypi_0
contourpy=1.0.5=py38h59b6b97_0
cryptography=39.0.1=py38h21b164f_0
cuda-cccl=12.1.55=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.1.55=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
d2l=0.17.6=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
efficientnet-pytorch=0.7.1=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.3=pypi_0
filelock=3.9.0=py38haa95532_0
flit-core=3.8.0=py38haa95532_0
fonttools=4.39.2=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
huggingface-hub=0.13.3=pypi_0
icu=58.2=ha925a31_3
idna=2.10=pypi_0
importlib-metadata=6.1.0=pypi_0
importlib-resources=5.12.0=pypi_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.22.0=pypi_0
ipython=8.11.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=py38haa95532_0
jpeg=9e=h2bbff1b_1
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.1.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.5.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.6=pypi_0
kiwisolver=1.4.4=py38hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.2.56=0
libcurand-dev=10.3.2.56=0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.5.1=pypi_0
matplotlib-base=3.7.1=py38hf11a4ad_1
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
mpmath=1.2.1=py38haa95532_0
munkres=1.1.4=py_0
nbclassic=0.5.3=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.10=pypi_0
nbformat=5.8.0=pypi_0
nest-asyncio=1.5.6=pypi_0
networkx=2.8.4=py38haa95532_1
notebook=6.5.3=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.21.5=pypi_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=py38haa95532_0
pandas=1.2.4=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py38hd77b12b_0
pip=23.0.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.1.1=pypi_0
ply=3.11=py38_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
psutil=5.9.4=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.14.0=pypi_0
pyopenssl=23.0.0=py38haa95532_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py38haa95532_0
python=3.8.16=h6244533_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-json-logger=2.0.7=pypi_0
pytorch=2.0.0=py3.8_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_3
pytorch-mutex=1.0=cuda
pytz=2023.2=pypi_0
pywin32=305=pypi_0
pywinpty=2.0.10=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtconsole=5.4.1=pypi_0
qtpy=2.3.0=pypi_0
qtwebkit=5.212=h3ad3cdb_4
requests=2.25.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
send2trash=1.8.0=pypi_0
setuptools=65.6.3=py38haa95532_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
soupsieve=2.4=pypi_0
sqlite=3.41.1=h2bbff1b_0
stack-data=0.6.2=pypi_0
sympy=1.11.1=py38haa95532_0
tabulate=0.9.0=pypi_0
terminado=0.17.1=pypi_0
timm=0.6.13=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=2.0.0=pypi_0
torchvision=0.15.0=pypi_0
tornado=6.2=py38h2bbff1b_0
tqdm=4.65.0=py38hd4e2768_0
traitlets=5.9.0=pypi_0
typing_extensions=4.4.0=py38haa95532_0
uri-template=1.2.0=pypi_0
urllib3=1.26.14=py38haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pypi_0
webcolors=1.12=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
wheel=0.38.4=py38haa95532_0
widgetsnbextension=4.0.6=pypi_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
xz=5.2.10=h8cc25b3_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
